Add vocabulary classes in a individual file
